+++ dashboard/package.json
@@ -38,7 +38,7 @@
     "gulp-rev-replace": "~0.4.2",
     "gulp-size": "~2.0.0",
     "gulp-sourcemaps": "~1.5.2",
-    "gulp-stylus": "~2.0.6",
+    "gulp-stylus": "2.5.0",
     "gulp-uglify": "~1.4.0",
     "gulp-useref": "~1.3.0",
     "gulp-util": "~3.0.6",
